+++ front/src/app/games/page.tsx
@@ -5,7 +5,7 @@ import { useLang } from '@/contexts/LanguageContext'
 import { SingleGame } from '@/components/Game/SingleGame'
 import { useAuth } from '@/contexts/AuthContext'
 import { Perms } from '@/interfaces/perms'
-import { SearchBar } from '@/components/SearchBar'
+import { SearchBar } from '@/components/Helpers/SearchBar'
 import { GameForm } from '@/components/Game/GameForm'
 import { Spinner } from '@/components/Messages/Spinner'
 import { useLoading } from '@/contexts/LoadingContext'
