+++ front/src/app/rentals/page.tsx
@@ -8,7 +8,7 @@ import ErrorField from '@/components/Messages/ErrorField'
 import SuccessField from '@/components/Messages/SuccessField'
 import { Spinner } from '@/components/Messages/Spinner'
 import { Perms } from '@/interfaces/perms'
-import { SearchBar } from '@/components/SearchBar'
+import { SearchBar } from '@/components/Helpers/SearchBar'
 import { RentalForm } from '@/components/Rental/RentalForm'
 import { RentalList } from '@/components/Rental/RentalList'
 import { useLoading } from '@/contexts/LoadingContext'
