+++ front/src/components/Helpers/DeleteDialog.tsx
