+++ front/src/components/Helpers/SearchBar.tsx
