--- front/src/components/LanguageToggle.tsx
@@ -1,56 +1,0 @@
-'use client'
-
-import { useLang } from '@/contexts/LanguageContext'
-
-/**
- * Component for toggling between different languages
- * @returns {JSX.Element} A button that toggles the application language
- */
-export function LanguageToggle() {
-  const { currentLang, setLanguage } = useLang()
-
-  return (
-    <div className="flex items-center space-x-2">
-      <button
-        onClick={() => setLanguage('pl')}
-        className={`px-2 py-1 rounded ${
-          currentLang === 'pl'
-            ? 'bg-blue-600 text-white'
-            : 'bg-gray-200 text-gray-700 hover:bg-gray-300 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600'
-        }`}
-      >
-        PL
-      </button>
-      <button
-        onClick={() => setLanguage('ua')}
-        className={`px-2 py-1 rounded ${
-          currentLang === 'ua'
-            ? 'bg-blue-600 text-white'
-            : 'bg-gray-200 text-gray-700 hover:bg-gray-300 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600'
-        }`}
-      >
-        UA
-      </button>
-      <button
-        onClick={() => setLanguage('en')}
-        className={`px-2 py-1 rounded ${
-          currentLang === 'en'
-            ? 'bg-blue-600 text-white'
-            : 'bg-gray-200 text-gray-700 hover:bg-gray-300 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600'
-        }`}
-      >
-        EN
-      </button>
-      <button
-        onClick={() => setLanguage('ja')}
-        className={`px-2 py-1 rounded ${
-          currentLang === 'ja'
-            ? 'bg-blue-600 text-white'
-            : 'bg-gray-200 text-gray-700 hover:bg-gray-300 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600'
-        }`}
-      >
-        日本語
-      </button>
-    </div>
-  )
-}
