--- front/src/lib/firebase/firebase.ts
@@ -1,33 +1,0 @@
-import { initializeApp, getApp, getApps } from 'firebase/app'
-import { getAuth, GoogleAuthProvider } from 'firebase/auth'
-import { getFirestore } from 'firebase/firestore'
-import { getStorage } from 'firebase/storage'
-
-const firebaseConfig = {
-  apiKey: "",
-  authDomain: "",
-  projectId: "",
-  storageBucket: "",
-  messagingSenderId: "",
-  appId: ""
-}
-
-// Initialize Firebase
-let app;
-try {
-  app = getApp();
-} catch {
-  app = initializeApp(firebaseConfig);
-}
-
-const auth = getAuth(app);
-const db = getFirestore(app);
-const storage = getStorage(app);
-
-// Configure Google Auth Provider settings
-const googleProvider = new GoogleAuthProvider();
-googleProvider.setCustomParameters({
-  prompt: 'select_account'
-});
-
-export { auth, db, storage, googleProvider }
