+++ front/src/types/gameContext.ts
@@ -6,13 +6,13 @@ export type SearchType = 'title' | 'category';
 export type GamesContextType = {
     games: IGame[]
     loading: boolean
+    searchQuery: string
+    searchType: SearchType
+    SearchedGames: IGame[]
     addGame: (game: Omit<IGame, 'id'>) => Promise<void>
     updateGame: (game: IGame) => Promise<void>
     deleteGame: (id: number) => void
     changeQuantity: (id: number, quantity: number) => void
-    searchType: SearchType
     setSearchType: (type: SearchType) => void
-    searchQuery: string
     setSearchQuery: (query: string) => void
-    SearchedGames: IGame[]
 }
